requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-storage
google-cloud-aiplatform
PyYAML
pydantic
python-dotenv